migrate all the trpc to server actions
trpc was 404ing so i was forced to migrate them which needed to be done anyway, but it's very annoying
