+++ apps/course-builder-web/package.json
@@ -54,7 +54,7 @@
 		"@mdx-js/react": "^3.0.0",
 		"@msgpack/msgpack": "3.0.0-beta2",
 		"@mux/mux-player-react": "^2.2.0",
-		"@next/mdx": "14.3.0-canary.5",
+		"@next/mdx": "14.3.0-canary.18",
 		"@pinecone-database/pinecone": "^2.1.0",
 		"@planetscale/database": "1.14.0",
 		"@radix-ui/colors": "^3.0.0",
@@ -120,7 +120,7 @@
 		"memoize-one": "^6.0.0",
 		"mjml": "^4.15.3",
 		"nanoid": "^5.0.2",
-		"next": "14.3.0-canary.5",
+		"next": "14.3.0-canary.18",
 		"next-auth": "5.0.0-beta.16",
 		"next-axiom": "^1.1.1",
 		"next-mdx-remote": "^4.4.1",
