+++ apps/course-builder-web/src/path-to-purchase/post-purchase-purchase-transfer.tsx
@@ -1,21 +1,17 @@
-'use client'
-
 import * as React from 'react'
-import { Transfer } from '@/purchase-transfer/purchase-transfer'
-import { api } from '@/trpc/react'
+import { revalidatePath } from 'next/cache'
+import { getPurchaseTransferForPurchaseId } from '@/purchase-transfer/purchase-transfer-actions'
+import { PurchaseTransferStatus } from '@/purchase-transfer/purchase-transfer-status'
 import { isEmpty } from 'lodash'
 
 export const PurchaseTransfer: React.FC<{
 	bulkCouponId?: string
 	purchase: { id: string; userId?: string | null }
-}> = ({ bulkCouponId, purchase }) => {
-	const { data: purchaseUserTransfers, refetch } =
-		api.purchaseUserTransfer.forPurchaseId.useQuery({
-			id: purchase.id,
-			sourceUserId: purchase.userId || undefined,
-		})
-
-	console.log({ purchaseUserTransfers })
+}> = async ({ bulkCouponId, purchase }) => {
+	const purchaseUserTransfers = await getPurchaseTransferForPurchaseId({
+		id: purchase.id,
+		sourceUserId: purchase.userId || undefined,
+	})
 
 	if (bulkCouponId) return null
 	if (isEmpty(purchaseUserTransfers)) return null
@@ -26,9 +22,12 @@ export const PurchaseTransfer: React.FC<{
 				Transfer this purchase to another email address
 			</h2>
 			{purchaseUserTransfers && (
-				<Transfer
+				<PurchaseTransferStatus
 					purchaseUserTransfers={purchaseUserTransfers}
-					refetch={refetch}
+					refetch={async () => {
+						'use server'
+						revalidatePath('/thanks/purchase')
+					}}
 				/>
 			)}
 		</div>
