+++ apps/course-builder-web/src/trpc/api/root.ts
@@ -1,7 +1,6 @@
 import { abilityRouter } from '@/trpc/api/routers/ability'
 import { imageResourceRouter } from '@/trpc/api/routers/imageResource'
 import { pricingRouter } from '@/trpc/api/routers/pricing'
-import { purchaseUserTransferRouter } from '@/trpc/api/routers/purchaseUserTransfer'
 import { usersRouter } from '@/trpc/api/routers/users'
 import { videoResourceRouter } from '@/trpc/api/routers/videoResource'
 import { createCallerFactory, createTRPCRouter } from '@/trpc/api/trpc'
@@ -16,7 +15,6 @@ export const appRouter = createTRPCRouter({
 	imageResources: imageResourceRouter,
 	users: usersRouter,
 	videoResources: videoResourceRouter,
-	purchaseUserTransfer: purchaseUserTransferRouter,
 	pricing: pricingRouter,
 })
